pkg-config: remove PKG_CONFIG and PKG_CONFIG_PATH from target neutrino
While commit 447a70e, 7535680 and 338f9d1 some working on pkg-config was
done. The target neutrino is the only one that use an extra setting of
PKG_CONFIG and PKG_CONFIG_PATH. That is'nt needed anymore because the BS
use its own builded pkg-config in it's own PATH.
This should  avoid problems when cross compiling.

Reviewed-by: Carsten Schönert <c.schoenert@t-online.de>