system-tools: add standalone wget
Dependencies are solved with build of e2fsprogs, because of libuuid.so
is required. Install of e2fsprogs is required. Not very nice but it works.
The busybox wget is not touched before installation of real wget.

Note: The "real" wget ist more larger than busybox version and
wget is needed by opkg-cl, but the busybox version
doesn't really working fine. It's recommended to add the "real" wget
to minimal-system-package or to flashimage for first install.
Installation directly on the stb is not so good idea.

[seife: fixed missing libuuid dependency in package, download URL,
        build failure wrt. pod2man]